requirements.txt
tensorflow>=2.10.0
opencv-python
flask
numpy
scikit-learn
matplotlib
pillow
flask>=2.0.0
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
Pillow>=8.0.0